chore: adopt ruff lint and format repo-wide

+++ tests/characterization/test_rendering.py
@@ -13,6 +13,7 @@ the legacy class; the observable text-only rendering fallback remains
 pinned. Display geometry mirrors the production wiring for the 4.2-inch
 panel: width 400 (``epd.height``) x height 300 (``epd.width``).
 """
+
 import logging
 from math import ceil
 from pathlib import Path
@@ -30,57 +31,59 @@ from quadstick_display.view import (
     ProfileRenderer,
 )
 
-RESOURCES = Path(__file__).resolve().parents[2] / 'resources'
-CSVS_DIR = RESOURCES / 'quadstick_csvs'
+RESOURCES = Path(__file__).resolve().parents[2] / "resources"
+CSVS_DIR = RESOURCES / "quadstick_csvs"
 
 DISPLAY_WIDTH = 400
 DISPLAY_HEIGHT = 300
 DISPLAY_SIZE = DisplaySize(width=DISPLAY_WIDTH, height=DISPLAY_HEIGHT)
 
-RADIUS = int(ceil(TEXT_SIZE / TEXT_SIZE_OFFSET))  # 7
+RADIUS = ceil(TEXT_SIZE / TEXT_SIZE_OFFSET)  # 7
 BUTTON_STRIDE = (RADIUS * 2) + 4  # 18
 
 # Verbatim copy of the production mouthpiece mapping table: three button
 # bits (left, center, right), then the is_puff bit, then the soft bit.
 EXPECTED_MP_BUTTONS = {
-    'mp_left_sip': (1, 0, 0, 0, 0),
-    'mp_left_puff': (1, 0, 0, 1, 0),
-    'mp_left_sip_soft': (1, 0, 0, 0, 1),
-    'mp_left_puff_soft': (1, 0, 0, 1, 1),
-    'mp_center_sip': (0, 1, 0, 0, 0),
-    'mp_center_puff': (0, 1, 0, 1, 0),
-    'mp_center_sip_soft': (0, 1, 0, 0, 1),
-    'mp_center_puff_soft': (0, 1, 0, 1, 1),
-    'mp_right_sip': (0, 0, 1, 0, 0),
-    'mp_right_puff': (0, 0, 1, 1, 0),
-    'mp_right_sip_soft': (0, 0, 1, 0, 1),
-    'mp_right_puff_soft': (0, 0, 1, 1, 1),
-    'mp_left_center_sip': (1, 1, 0, 0, 0),
-    'mp_left_center_puff': (1, 1, 0, 1, 0),
-    'mp_left_center_sip_soft': (1, 1, 0, 0, 1),
-    'mp_left_center_puff_soft': (1, 1, 0, 1, 1),
-    'mp_right_center_sip': (0, 1, 1, 0, 0),
-    'mp_right_center_puff': (0, 1, 1, 1, 0),
-    'mp_right_center_sip_soft': (0, 1, 1, 0, 1),
-    'mp_right_center_puff_soft': (0, 1, 1, 1, 1),
-    'mp_triple_sip': (1, 1, 1, 0, 0),
-    'mp_triple_puff': (1, 1, 1, 1, 0),
-    'mp_triple_sip_soft': (1, 1, 1, 0, 1),
-    'mp_triple_puff_soft': (1, 1, 1, 1, 1),
+    "mp_left_sip": (1, 0, 0, 0, 0),
+    "mp_left_puff": (1, 0, 0, 1, 0),
+    "mp_left_sip_soft": (1, 0, 0, 0, 1),
+    "mp_left_puff_soft": (1, 0, 0, 1, 1),
+    "mp_center_sip": (0, 1, 0, 0, 0),
+    "mp_center_puff": (0, 1, 0, 1, 0),
+    "mp_center_sip_soft": (0, 1, 0, 0, 1),
+    "mp_center_puff_soft": (0, 1, 0, 1, 1),
+    "mp_right_sip": (0, 0, 1, 0, 0),
+    "mp_right_puff": (0, 0, 1, 1, 0),
+    "mp_right_sip_soft": (0, 0, 1, 0, 1),
+    "mp_right_puff_soft": (0, 0, 1, 1, 1),
+    "mp_left_center_sip": (1, 1, 0, 0, 0),
+    "mp_left_center_puff": (1, 1, 0, 1, 0),
+    "mp_left_center_sip_soft": (1, 1, 0, 0, 1),
+    "mp_left_center_puff_soft": (1, 1, 0, 1, 1),
+    "mp_right_center_sip": (0, 1, 1, 0, 0),
+    "mp_right_center_puff": (0, 1, 1, 1, 0),
+    "mp_right_center_sip_soft": (0, 1, 1, 0, 1),
+    "mp_right_center_puff_soft": (0, 1, 1, 1, 1),
+    "mp_triple_sip": (1, 1, 1, 0, 0),
+    "mp_triple_puff": (1, 1, 1, 1, 0),
+    "mp_triple_sip_soft": (1, 1, 1, 0, 1),
+    "mp_triple_puff_soft": (1, 1, 1, 1, 1),
 }
 
-ONE_ROW = [('kb_space', 'mp_center_sip')]
+ONE_ROW = [("kb_space", "mp_center_sip")]
 
 
-@pytest.fixture(scope='module')
+@pytest.fixture(scope="module")
 def renderer():
     return ProfileRenderer(RESOURCES)
 
 
 def make_profile(rows):
     return Profile(
-        name='test',
-        bindings=tuple(Binding(command, quadstick_input) for command, quadstick_input in rows),
+        name="test",
+        bindings=tuple(
+            Binding(command, quadstick_input) for command, quadstick_input in rows
+        ),
     )
 
 
@@ -97,13 +100,15 @@ def mouthpiece_origin(renderer, rows, row_index=0):
     """Top-left reference for the mouthpiece cluster of a rendered row."""
     x2 = column_split(renderer, rows)
     x0 = int(x2 + TEXT_SIZE)
-    y_center = int(float(row_index) * TEXT_SIZE + TEXT_SIZE - (TEXT_SIZE // TEXT_SIZE_OFFSET))
+    y_center = int(
+        float(row_index) * TEXT_SIZE + TEXT_SIZE - (TEXT_SIZE // TEXT_SIZE_OFFSET)
+    )
     return x0, y_center
 
 
 def dark_rows_in_band(image, x0, x1):
     """Row indices containing a dark pixel within columns [x0, x1)."""
-    gray = image.convert('L')
+    gray = image.convert("L")
     width, height = gray.size
     rows = set()
     for x in range(max(0, x0), min(x1, width)):
@@ -129,17 +134,17 @@ class TestMouthpieceTable:
 
 class TestImageBasics:
     def test_planes_mode_and_dimensions(self, renderer):
-        frame = render(renderer, [('kb_w', 'up'), ('kb_s', 'down')])
+        frame = render(renderer, [("kb_w", "up"), ("kb_s", "down")])
         assert frame.black is not frame.red
         for plane in (frame.black, frame.red):
-            assert plane.mode == IMAGE_MODE == '1'
+            assert plane.mode == IMAGE_MODE == "1"
             assert plane.size == (DISPLAY_WIDTH, DISPLAY_HEIGHT)
             # Nothing reaches the bottom-right corner at this content size.
             assert plane.getpixel((DISPLAY_WIDTH - 1, DISPLAY_HEIGHT - 1)) == 255
         assert 0 in frame.black.getdata()  # content was drawn
 
     def test_separator_is_split_across_planes(self, renderer):
-        rows = [('kb_w', 'up'), ('kb_s', 'down')]
+        rows = [("kb_w", "up"), ("kb_s", "down")]
         frame = render(renderer, rows)
         for row_index in range(2):
             y_line = row_index * TEXT_SIZE + 2
@@ -158,9 +163,9 @@ class TestImageBasics:
 
 
 class TestMouthpieceRendering:
-    @pytest.mark.parametrize('button', sorted(EXPECTED_MP_BUTTONS))
+    @pytest.mark.parametrize("button", sorted(EXPECTED_MP_BUTTONS))
     def test_button_glyphs_match_mapping_table(self, renderer, button):
-        rows = [('kb_space', button)]
+        rows = [("kb_space", button)]
         frame = render(renderer, rows)
         config = EXPECTED_MP_BUTTONS[button]
         x0, y_center = mouthpiece_origin(renderer, rows)
@@ -191,7 +196,7 @@ class TestMouthpieceRendering:
     def test_text_only_fallback_for_non_mouthpiece_outputs(self, renderer):
         # Outputs that are not mp_* buttons (e.g. 'right_puff' -> 'R.PUFF')
         # render as red-plane text: no circles, no black-plane icon.
-        rows = [('kb_space', 'right_puff')]
+        rows = [("kb_space", "right_puff")]
         frame = render(renderer, rows)
         x0, _ = mouthpiece_origin(renderer, rows)
 
@@ -202,7 +207,7 @@ class TestMouthpieceRendering:
 
 class TestArrowOrientation:
     def test_icons_are_resized_to_text_height(self, renderer):
-        raw = Image.open(RESOURCES / 'images' / 'airflow_arrow_ltr.png')
+        raw = Image.open(RESOURCES / "images" / "airflow_arrow_ltr.png")
         expected_width = int((raw.width / raw.height) * TEXT_SIZE)
         assert renderer.arrow_right_icon.size == (expected_width, TEXT_SIZE) == (28, 18)
         assert renderer.arrow_left_icon.size == renderer.arrow_right_icon.size
@@ -210,7 +215,7 @@ class TestArrowOrientation:
     def test_left_arrow_is_horizontal_flip_of_right_arrow(self, renderer):
         flipped = renderer.arrow_right_icon.transpose(Image.Transpose.FLIP_LEFT_RIGHT)
         diff = ImageChops.difference(
-            renderer.arrow_left_icon.convert('L'), flipped.convert('L')
+            renderer.arrow_left_icon.convert("L"), flipped.convert("L")
         )
         assert diff.getbbox() is None
 
@@ -219,20 +224,20 @@ class TestArrowOrientation:
         # centered tip is on the right edge; the wide tail mass is on the
         # left. So the unflipped icon points right and the flipped one
         # points left.
-        raw = Image.open(RESOURCES / 'images' / 'airflow_arrow_ltr.png')
+        raw = Image.open(RESOURCES / "images" / "airflow_arrow_ltr.png")
         left = dark_rows_in_band(raw, 0, 10)
         right = dark_rows_in_band(raw, raw.width - 10, raw.width)
         assert len(right) < len(left)
         assert min(right) < raw.height // 2 < max(right)
 
     def test_sip_points_left_and_puff_points_right(self, renderer):
-        frame_sip = render(renderer, [('kb_space', 'mp_center_sip')])
-        frame_puff = render(renderer, [('kb_space', 'mp_center_puff')])
+        frame_sip = render(renderer, [("kb_space", "mp_center_sip")])
+        frame_puff = render(renderer, [("kb_space", "mp_center_puff")])
 
         # Same command text and same circles: the black planes differ only
         # where the arrow icon was pasted.
         diff = ImageChops.difference(
-            frame_sip.black.convert('L'), frame_puff.black.convert('L')
+            frame_sip.black.convert("L"), frame_puff.black.convert("L")
         )
         bbox = diff.getbbox()
         assert bbox is not None
@@ -242,8 +247,8 @@ class TestArrowOrientation:
 
         # The two pasted icons are exact horizontal mirrors.
         mirror = ImageChops.difference(
-            region_sip.convert('L'),
-            region_puff.transpose(Image.Transpose.FLIP_LEFT_RIGHT).convert('L'),
+            region_sip.convert("L"),
+            region_puff.transpose(Image.Transpose.FLIP_LEFT_RIGHT).convert("L"),
         )
         assert mirror.getbbox() is None
 
@@ -263,30 +268,28 @@ class TestOverflowResizing:
     def test_oversized_content_is_downscaled_to_display_size(self, renderer, caplog):
         # 24 rows * 18px = 432px > 300px display height: the image is drawn
         # tall and then resized back with BOX resampling.
-        profile = ProfileStore(CSVS_DIR).load('alan_wake_II.csv')
+        profile = ProfileStore(CSVS_DIR).load("alan_wake_II.csv")
         with caplog.at_level(logging.INFO):
             frame = renderer.render(profile, DISPLAY_SIZE)
         for plane in (frame.black, frame.red):
             assert plane.size == (DISPLAY_WIDTH, DISPLAY_HEIGHT)
-            assert plane.mode == '1'
-        assert 'Resizing image to 400x432' in caplog.text
-        assert 'Resizing image back to the display size 400x300' in caplog.text
+            assert plane.mode == "1"
+        assert "Resizing image to 400x432" in caplog.text
+        assert "Resizing image back to the display size 400x300" in caplog.text
 
     def test_content_that_fits_is_not_resized(self, renderer, caplog):
         # 16 rows * 18px = 288px <= 300px: the image is drawn at display
         # size and no resize happens.
-        profile = ProfileStore(CSVS_DIR).load('ad_infinitum.csv')
+        profile = ProfileStore(CSVS_DIR).load("ad_infinitum.csv")
         with caplog.at_level(logging.INFO):
             frame = renderer.render(profile, DISPLAY_SIZE)
         for plane in (frame.black, frame.red):
             assert plane.size == (DISPLAY_WIDTH, DISPLAY_HEIGHT)
-        assert 'Resizing image' not in caplog.text
+        assert "Resizing image" not in caplog.text
 
     def test_taller_display_yields_taller_image(self, renderer, caplog):
-        profile = ProfileStore(CSVS_DIR).load('alan_wake_II.csv')
+        profile = ProfileStore(CSVS_DIR).load("alan_wake_II.csv")
         with caplog.at_level(logging.INFO):
-            frame = renderer.render(
-                profile, DisplaySize(width=400, height=500)
-            )
+            frame = renderer.render(profile, DisplaySize(width=400, height=500))
         assert frame.black.size == (400, 500)
-        assert 'Resizing image' not in caplog.text
+        assert "Resizing image" not in caplog.text
